前処理・fmt処理・lint処理を別ファイルに分離
