+++ src/linter/linter.py
@@ -1,0 +1,2 @@
+def lint():
+  print("lint")
